#59190 UPD: report copy fixes

See merge request dev/oez!447

+++ www/local/modules/kelnik.report/lib/model/reportfieldstable.php
@@ -203,22 +203,22 @@ class ReportFieldsTable extends DataManager
                             [
                                 'id' => 'jobs-plan-all',
                                 'title' => Loc::getMessage('KELNIK_REPORT_FIELDS_1-2-1'),
-                                'type' => 'float'
+                                'type' => 'int'
                             ],
                             [
                                 'id' => 'jobs-plan-year',
                                 'title' => Loc::getMessage('KELNIK_REPORT_FIELDS_1-2-2'),
-                                'type' => 'float'
+                                'type' => 'int'
                             ],
                             [
                                 'id' => 'jobs-actual-all',
                                 'title' => Loc::getMessage('KELNIK_REPORT_FIELDS_1-2-3'),
-                                'type' => 'float'
+                                'type' => 'int'
                             ],
                             [
                                 'id' => 'jobs-actual-year',
                                 'title' => Loc::getMessage('KELNIK_REPORT_FIELDS_1-2-4'),
-                                'type' => 'float'
+                                'type' => 'int'
                             ]
                         ]
                     ],
